Extract TUI statusline module
