+++ crates/warp_tui/src/terminal_session_view.rs
@@ -8,13 +8,12 @@ use std::time::Duration;
 
 use ai::LLMProvider;
 use async_channel::Sender;
-use chrono::{Local, NaiveDateTime};
 use instant::Instant;
 use parking_lot::FairMutex;
 use warp::editor::{CodeEditorModel, CodeEditorModelEvent};
 use warp::settings::{
-    AISettings, AISettingsChangedEvent, AppEditorSettings, TuiStatuslineConfig, TuiStatuslineItem,
-    TuiTheme, TuiThemeSettings, TuiVoiceSettings,
+    AISettings, AISettingsChangedEvent, AppEditorSettings, TuiStatuslineConfig, TuiTheme,
+    TuiThemeSettings, TuiVoiceSettings,
 };
 use warp::tui_export::{
     AIAgentActionId, AIAgentActionResultType, AIAgentContext, AIAgentExchangeId,
@@ -25,9 +24,9 @@ use warp::tui_export::{
     BlocklistAIHistoryEvent, BlocklistAIHistoryModel, BlocklistAIInputModel, CLISubagentController,
     CLISubagentEvent, CLISubagentTarget, COMMAND_REGISTRY, CancellationReason, ChangelogModel,
     ChangelogRequestType, CloudConversationData, CommandExecutionSource, ConversationFileExport,
-    ConversationSelection, ConversationSelectionHandle, ConversationUsageTotals,
-    ExecuteCommandEvent, GetRelevantFilesController, GitHubRepoModel, GitRepoModels,
-    GitRepoStatusModel, GitStatusMetadata, LLMId, LLMPreferences, LLMPreferencesEvent,
+    ConversationSelection, ConversationSelectionHandle, ExecuteCommandEvent,
+    GetRelevantFilesController, GitHubRepoModel, GitRepoStatusModel, LLMId, LLMPreferences,
+    LLMPreferencesEvent,
     LOCAL_SKILLS_REMOTE_EXECUTION_ERROR_MESSAGE, LinkedWorkflowData, ModelEvent,
     ParsedSlashCommandInput, PersistenceWriter, PtyIntent, PtyIntentEvent, QueuedQueryEvent,
     QueuedQueryModel, RepoDetectionSessionType, RepoDetectionSource, ServerConversationToken,
@@ -121,15 +120,12 @@ use crate::terminal_use::{
     terminal_use_conversation_to_resume, terminal_use_interrupt_action, tui_input_target,
 };
 use crate::transcript_view::{TuiTranscriptView, TuiTranscriptViewEvent};
-use crate::transient_hint::{TransientHint, TransientHintTone};
+use crate::transient_hint::TransientHint;
 use crate::tui_ask_question_view::TuiAskQuestionView;
 use crate::tui_builder::TuiUiBuilder;
 use crate::tui_cli_subagent_view::{HAND_BACK_KEY_BINDING, TuiCLISubagentView};
 use crate::tui_permission_prompt::TuiPermissionPrompt;
-use crate::ui::{
-    abbreviate_home_prefix, compact_footer_path, conversation_restore_failed,
-    conversation_restoring,
-};
+use crate::ui::{abbreviate_home_prefix, conversation_restore_failed, conversation_restoring};
 use crate::usage::UsageToggle;
 use crate::voice_input::{
     TuiVoiceInputEvent, TuiVoiceInputState, VoiceInputStartSource, configured_hold_key,
@@ -149,6 +145,7 @@ mod input_detection;
 mod shortcuts;
 pub(crate) mod state;
 mod status_menu;
+mod statusline;
 mod todo_menu;
 use self::completions::CompletionRequestState;
 use self::input_detection::InputDetectionState;
@@ -162,7 +159,6 @@ const INLINE_MENU_TOP_PADDING_ROWS: u16 = 1;
 const MAX_READ_ONLY_MENU_ROWS: u16 = 10;
 const MAX_INPUT_TEXT_ROWS: u16 = 6;
 const AUTO_APPROVE_FEEDBACK_DURATION: Duration = Duration::from_secs(3);
-const STATUSLINE_DATETIME_REPAINT_INTERVAL: Duration = Duration::from_secs(60);
 const VOICE_INPUT_BORDER_REPAINT_INTERVAL: Duration = Duration::from_millis(33);
 
 /// The footer hint shown while the ctrl-c exit confirmation is armed.
@@ -334,49 +330,6 @@ const COST_EMPTY_CONVERSATION_HINT: &str = "Cannot show conversation cost: conve
 const COST_CONVERSATION_IN_PROGRESS_HINT: &str =
     "Cannot show conversation cost: conversation is in progress";
 
-struct FooterHint<'a> {
-    text: &'a str,
-    style: FooterHintStyle,
-}
-
-enum FooterHintStyle {
-    Muted,
-    Success,
-    Error,
-    VoiceInput,
-}
-
-impl<'a> FooterHint<'a> {
-    fn muted(text: &'a str) -> Self {
-        Self {
-            text,
-            style: FooterHintStyle::Muted,
-        }
-    }
-
-    fn voice_input(text: &'a str) -> Self {
-        Self {
-            text,
-            style: FooterHintStyle::VoiceInput,
-        }
-    }
-
-    fn render(self, builder: &TuiUiBuilder) -> TuiFlex {
-        let style = match self.style {
-            FooterHintStyle::Muted => builder.muted_text_style(),
-            FooterHintStyle::Success => builder.success_glyph_style(),
-            FooterHintStyle::Error => builder.error_text_style(),
-            FooterHintStyle::VoiceInput => builder.voice_input_status_style(),
-        };
-        TuiFlex::row().child(
-            TuiText::new(self.text)
-                .with_style(style)
-                .truncate()
-                .finish(),
-        )
-    }
-}
-
 fn log_bundle_success_message(path: &Path) -> String {
     format!("Log bundle saved to {}", path.display())
 }
@@ -430,49 +383,6 @@ fn provider_api_key_shell_command(
 fn raw_prompt_if_not_blank(input: &str) -> Option<&str> {
     (!input.trim().is_empty()).then_some(input)
 }
-#[derive(Debug, PartialEq, Eq)]
-struct ContextWindowUsage {
-    bar: String,
-    percentage_remaining: u8,
-    warning: bool,
-}
-fn format_context_window_usage(usage: f32) -> ContextWindowUsage {
-    const BAR_WIDTH: usize = 4;
-
-    let remaining = (1.0 - usage).clamp(0.0, 1.0);
-    let percentage_remaining = (remaining * 100.0).round() as u8;
-    let filled = ((remaining * BAR_WIDTH as f32) + f32::EPSILON).floor() as usize;
-    ContextWindowUsage {
-        bar: format!("{}{}", "█".repeat(filled), "░".repeat(BAR_WIDTH - filled)),
-        percentage_remaining,
-        warning: percentage_remaining <= 25,
-    }
-}
-fn format_statusline_date(now: NaiveDateTime) -> String {
-    now.format("%B %-d, %Y").to_string()
-}
-fn format_statusline_time_12_hour(now: NaiveDateTime) -> String {
-    now.format("%-I:%M%P").to_string()
-}
-fn format_statusline_time_24_hour(now: NaiveDateTime) -> String {
-    now.format("%H:%M").to_string()
-}
-fn render_statusline_datetime(
-    formatter: fn(NaiveDateTime) -> String,
-    style: TuiStyle,
-) -> Box<dyn TuiElement> {
-    TuiAnimated::new(STATUSLINE_DATETIME_REPAINT_INTERVAL, move || {
-        TuiText::new(formatter(Local::now().naive_local()))
-            .with_style(style)
-            .truncate()
-            .finish()
-    })
-    .finish()
-}
-fn format_todo_progress(completed: usize, total: usize, finished: bool) -> String {
-    let marker = if finished { "✓" } else { "❒" };
-    format!("{marker} {completed}/{total}")
-}
 fn cost_command_unavailable_hint(
     selected_conversation: Option<(bool, bool)>,
 ) -> Option<&'static str> {
@@ -511,179 +421,6 @@ fn bordered_input(
         .finish()
 }
 
-/// One resolved item in the footer's configured presentation order.
-enum FooterSegment {
-    ShellMode,
-    AutoApproveIndicator(Box<dyn TuiElement>),
-    /// Vim mode indicator (NOR/INS/VIS/V-L/REP), driven by the VimModeIndicator statusline item.
-    VimIndicator(&'static str),
-    Model(Box<dyn TuiElement>),
-    WorkingDirectory(String),
-    GitBranch(String),
-    CreditUsage(Box<dyn TuiElement>),
-    ContextWindowUsage(ContextWindowUsage),
-    GitDiff {
-        files_changed: usize,
-        additions: usize,
-        deletions: usize,
-    },
-    GitBranchStatus(Box<dyn TuiElement>),
-    GitHubPullRequest(Box<dyn TuiElement>),
-    DateTime(Box<dyn TuiElement>),
-    AgentTodoList(Box<dyn TuiElement>),
-    VoiceInput(Box<dyn TuiElement>),
-}
-
-impl FooterSegment {
-    fn separator_to(&self, next: &Self) -> &'static str {
-        match (self, next) {
-            (
-                Self::ShellMode | Self::Model(_) | Self::VimIndicator(_),
-                Self::WorkingDirectory(_),
-            ) => " ",
-            (Self::WorkingDirectory(_), Self::GitBranch(_) | Self::GitBranchStatus(_)) => " ",
-            (
-                Self::WorkingDirectory(_) | Self::GitBranch(_),
-                Self::WorkingDirectory(_) | Self::GitBranch(_),
-            )
-            | (Self::DateTime(_), Self::DateTime(_))
-            | (Self::ShellMode, _)
-            | (_, Self::ShellMode) => " • ",
-            (
-                Self::AutoApproveIndicator(_)
-                | Self::VimIndicator(_)
-                | Self::Model(_)
-                | Self::WorkingDirectory(_)
-                | Self::GitBranch(_)
-                | Self::CreditUsage(_)
-                | Self::ContextWindowUsage(_)
-                | Self::GitDiff { .. }
-                | Self::GitBranchStatus(_)
-                | Self::GitHubPullRequest(_)
-                | Self::DateTime(_)
-                | Self::AgentTodoList(_)
-                | Self::VoiceInput(_),
-                Self::AutoApproveIndicator(_)
-                | Self::VimIndicator(_)
-                | Self::Model(_)
-                | Self::WorkingDirectory(_)
-                | Self::GitBranch(_)
-                | Self::CreditUsage(_)
-                | Self::ContextWindowUsage(_)
-                | Self::GitDiff { .. }
-                | Self::GitBranchStatus(_)
-                | Self::GitHubPullRequest(_)
-                | Self::DateTime(_)
-                | Self::AgentTodoList(_)
-                | Self::VoiceInput(_),
-            ) => " | ",
-        }
-    }
-}
-
-/// Resolved segments for the footer's left-aligned status row.
-struct FooterSegments {
-    ordered: Vec<FooterSegment>,
-}
-/// Builds the status row from resolved segments. Working directory follows a
-/// leading shell-mode label with a plain space; a branch or composite branch
-/// status owns its `⊢` glyph and follows working directory with one space.
-/// Items in different Figma groups use ` | `; other adjacent pairs use ` • `.
-/// The first item never receives a separator.
-fn render_status_footer_row(segments: FooterSegments, builder: &TuiUiBuilder) -> TuiFlex {
-    let muted = builder.muted_text_style();
-    let mut row = TuiFlex::row();
-    let mut segments = segments.ordered.into_iter().peekable();
-    while let Some(segment) = segments.next() {
-        let separator = segments.peek().map(|next| segment.separator_to(next));
-        match segment {
-            FooterSegment::ShellMode => {
-                row = row.child(
-                    TuiText::new(SHELL_MODE_HINT)
-                        .with_style(builder.shell_command_accent_style())
-                        .truncate()
-                        .finish(),
-                );
-            }
-            FooterSegment::VimIndicator(label) => {
-                row = row.child(
-                    TuiText::new(label)
-                        .with_style(builder.accent_border_style())
-                        .truncate()
-                        .finish(),
-                );
-            }
-            FooterSegment::AutoApproveIndicator(element)
-            | FooterSegment::Model(element)
-            | FooterSegment::CreditUsage(element)
-            | FooterSegment::GitBranchStatus(element)
-            | FooterSegment::GitHubPullRequest(element)
-            | FooterSegment::DateTime(element)
-            | FooterSegment::AgentTodoList(element)
-            | FooterSegment::VoiceInput(element) => {
-                row = row.child(element);
-            }
-            FooterSegment::WorkingDirectory(cwd) => {
-                row = row.child(TuiText::new(cwd).with_style(muted).truncate().finish());
-            }
-            FooterSegment::GitBranch(branch) => {
-                row = row.child(
-                    TuiText::new(format!("⊢ {branch}"))
-                        .with_style(muted)
-                        .truncate()
-                        .finish(),
-                );
-            }
-            FooterSegment::ContextWindowUsage(usage) => {
-                let value_style = if usage.warning {
-                    builder.attention_glyph_style()
-                } else {
-                    builder.primary_text_style()
-                };
-                row = row.child(
-                    TuiText::from_spans([
-                        (
-                            format!("{} {}% ", usage.bar, usage.percentage_remaining),
-                            value_style,
-                        ),
-                        ("context remaining".to_owned(), muted),
-                    ])
-                    .truncate()
-                    .finish(),
-                );
-            }
-            FooterSegment::GitDiff {
-                files_changed,
-                additions,
-                deletions,
-            } => {
-                let mut spans = vec![(format!("☰ {files_changed}"), muted)];
-                if additions > 0 || deletions > 0 {
-                    spans.push((" •".to_owned(), muted));
-                }
-                if additions > 0 {
-                    spans.push((format!(" +{additions}"), builder.diff_added_style()));
-                }
-                if deletions > 0 {
-                    spans.push((" ".to_owned(), muted));
-                    spans.push((format!("-{deletions}"), builder.diff_removed_style()));
-                }
-                row = row.child(TuiText::from_spans(spans).truncate().finish());
-            }
-        }
-        if let Some(separator) = separator {
-            row = row.child(
-                TuiText::new(separator)
-                    .with_style(muted)
-                    .truncate()
-                    .finish(),
-            );
-        }
-    }
-
-    row
-}
-
 fn mcp_primary_action_hint(action: TuiMcpAction) -> Option<&'static str> {
     match action {
         TuiMcpAction::Start(_) => Some("to start"),
@@ -720,43 +457,6 @@ fn render_mcp_menu_footer(
         (" to close".to_owned(), builder.muted_text_style()),
     ]);
     TuiFlex::row().child(TuiText::from_spans(spans).truncate().finish())
-}
-fn render_git_branch_status(
-    branch: &str,
-    rebased: bool,
-    ahead: Option<String>,
-    behind: Option<String>,
-    builder: &TuiUiBuilder,
-) -> Box<dyn TuiElement> {
-    let muted = builder.muted_text_style();
-    let accent = builder.accent_text_style();
-    let has_ahead = ahead.is_some();
-    let has_behind = behind.is_some();
-    let mut spans = vec![(format!("⊢ {branch}"), muted)];
-    if rebased || has_ahead || has_behind {
-        spans.push((" • ".to_owned(), muted));
-    }
-    if rebased {
-        spans.push(("⇅".to_owned(), accent));
-    } else {
-        if let Some(ahead) = ahead {
-            spans.push(("↑".to_owned(), accent));
-            spans.push((
-                format!("{ahead}{}", if has_behind { " " } else { "" }),
-                muted,
-            ));
-        }
-        if let Some(behind) = behind {
-            spans.push(("↓".to_owned(), accent));
-            spans.push((behind, muted));
-        }
-    }
-    TuiText::from_spans(spans).truncate().finish()
-}
-
-fn should_render_plain_git_branch(config: &TuiStatuslineConfig) -> bool {
-    config.is_enabled(TuiStatuslineItem::GitBranch)
-        && !config.is_enabled(TuiStatuslineItem::GitBranchStatus)
 }
 /// Entry point that requested conversation restoration.
 #[derive(Clone, Copy, Debug)]
@@ -3381,431 +3081,6 @@ impl TuiTerminalSessionView {
         render_warping_indicator_row(label, elapsed, auto_approve, ctx)
     }
 
-    /// Selects the single message that replaces the normal footer, preserving
-    /// the priority order between competing session states.
-    fn footer_hint(
-        &self,
-        voice_statusline_visible: bool,
-        ctx: &AppContext,
-    ) -> Option<FooterHint<'_>> {
-        if self.exit_confirmation.is_armed() {
-            // When the kill-child window is armed, show the child-specific hint
-            // so the user knows the next ctrl-c will kill the child agent rather
-            // than exiting the whole TUI.
-            if self.child_kill_armed_conversation.is_some() {
-                return Some(FooterHint::muted(CTRL_C_KILL_CHILD_HINT));
-            }
-            return Some(FooterHint::muted(CTRL_C_EXIT_HINT));
-        }
-        if matches!(
-            &self.conversation_restore_state,
-            ConversationRestoreState::Loading {
-                origin: TuiConversationRestoreOrigin::ConversationList,
-                ..
-            }
-        ) {
-            return Some(FooterHint::muted(LOADING_CONVERSATION_HINT));
-        }
-        if let Some((text, tone)) = self.transient_hint.current() {
-            let style = match tone {
-                TransientHintTone::Muted => FooterHintStyle::Muted,
-                TransientHintTone::Success => FooterHintStyle::Success,
-                TransientHintTone::Error => FooterHintStyle::Error,
-            };
-            return Some(FooterHint { text, style });
-        }
-        if self
-            .session_state(ctx)
-            .is_ok_and(|state| state.agent_is_tagged_in())
-        {
-            return Some(FooterHint::muted(RUNNING_COMMAND_DETACH_HINT));
-        }
-        if voice_statusline_visible {
-            return None;
-        }
-        match self.input_view.as_ref(ctx).voice_state(ctx) {
-            TuiVoiceInputState::Listening => {
-                let hint = if self.input_view.as_ref(ctx).voice_hold_key(ctx).is_some() {
-                    "listening to voice input... · release key to stop"
-                } else {
-                    "listening to voice input... · esc or enter to stop"
-                };
-                return Some(FooterHint::voice_input(hint));
-            }
-            TuiVoiceInputState::Transcribing => {
-                return Some(FooterHint::voice_input("Transcribing... · esc to cancel"));
-            }
-            TuiVoiceInputState::Idle => {}
-        }
-        None
-    }
-
-    /// Builds the configured statusline under the input box. Normal mode uses
-    /// the persisted item order and visibility; shell mode always leads with
-    /// its mode label and resolves configured shell-relevant metadata. A
-    /// replacing hint — the ctrl-c exit confirmation while armed, the
-    /// conversation-list loading hint, an active transient notice, or the
-    /// interrupt hint for a manually attached running command — occupies the
-    /// whole row instead. The open MCP menu similarly replaces the statusline
-    /// with its controls. An empty resolved configuration consumes no row.
-    fn render_footer(&self, ctx: &AppContext) -> TuiFlex {
-        let builder = TuiUiBuilder::from_app(ctx);
-        let shell_mode = self.is_shell_mode(ctx);
-        let config = AISettings::as_ref(ctx).tui_statusline.normalized();
-        let voice_statusline_visible = config.is_enabled(TuiStatuslineItem::VoiceInput)
-            && self.voice_statusline_is_available(shell_mode, ctx);
-        if let Some(hint) = self.footer_hint(voice_statusline_visible, ctx) {
-            return hint.render(&builder);
-        }
-        if self.mcp_menu.as_ref(ctx).is_open(ctx) {
-            let menu = self.mcp_menu.as_ref(ctx);
-            return render_mcp_menu_footer(
-                &builder,
-                menu.selected_primary_action(ctx),
-                menu.can_log_out_selected(ctx),
-            );
-        }
-        let git_metadata = self.git_status_metadata(ctx);
-        let mut ordered = Vec::new();
-        if shell_mode {
-            ordered.push(FooterSegment::ShellMode);
-        }
-        for item in config.order.iter().copied() {
-            if !config.is_enabled(item) {
-                continue;
-            }
-            let segment = match item {
-                TuiStatuslineItem::AutoApprove => (!shell_mode).then(|| {
-                    FooterSegment::AutoApproveIndicator(
-                        self.render_auto_approve_statusline(&builder, ctx),
-                    )
-                }),
-                TuiStatuslineItem::VimModeIndicator => self
-                    .vim_mode_indicator(ctx)
-                    .map(FooterSegment::VimIndicator),
-                TuiStatuslineItem::Model => (!shell_mode).then(|| {
-                    let model_name = LLMPreferences::as_ref(ctx)
-                        .get_active_base_model(ctx, Some(self.terminal_surface_id))
-                        .display_name
-                        .clone();
-                    let model_label_hovered = self
-                        .model_label_hover
-                        .lock()
-                        .is_ok_and(|state| state.is_hovered());
-                    let model_label_style = if model_label_hovered {
-                        builder.primary_text_style()
-                    } else {
-                        builder.muted_text_style()
-                    };
-                    FooterSegment::Model(
-                        TuiHoverable::new(
-                            self.model_label_hover.clone(),
-                            TuiText::new(model_name)
-                                .with_style(model_label_style)
-                                .truncate()
-                                .finish(),
-                        )
-                        .on_click(|event_ctx, _| {
-                            event_ctx
-                                .dispatch_typed_action(TuiTerminalSessionAction::ToggleModelMenu);
-                        })
-                        .finish(),
-                    )
-                }),
-                TuiStatuslineItem::GitHubPullRequest => (!shell_mode)
-                    .then_some(self.github_repo.as_ref())
-                    .flatten()
-                    .and_then(|repo| repo.as_ref(ctx).pr_info(ctx))
-                    .map(|pr| {
-                        let url = pr.url.clone();
-                        FooterSegment::GitHubPullRequest(self.github_pr_link.render(
-                            format!("PR #{}", pr.number),
-                            ctx,
-                            move |event_ctx, _| {
-                                event_ctx.dispatch_typed_action(TuiTerminalSessionAction::OpenUrl(
-                                    url.clone(),
-                                ));
-                            },
-                        ))
-                    }),
-                TuiStatuslineItem::WorkingDirectory => self
-                    .current_working_directory(ctx)
-                    .map(|cwd| FooterSegment::WorkingDirectory(compact_footer_path(&cwd))),
-                TuiStatuslineItem::GitBranch => should_render_plain_git_branch(&config)
-                    .then(|| {
-                        git_metadata.map(|metadata| {
-                            FooterSegment::GitBranch(metadata.current_branch_name.clone())
-                        })
-                    })
-                    .flatten(),
-                TuiStatuslineItem::GitBranchStatus => git_metadata.map(|metadata| {
-                    let tracking = &metadata.branch_tracking_status;
-                    FooterSegment::GitBranchStatus(render_git_branch_status(
-                        &metadata.current_branch_name,
-                        tracking.is_rebased(),
-                        tracking.ahead_display_count(),
-                        tracking.behind_display_count(),
-                        &builder,
-                    ))
-                }),
-                TuiStatuslineItem::GitDiffStatus => git_metadata.and_then(|metadata| {
-                    let stats = metadata.stats_against_head;
-                    (stats.files_changed > 0).then_some(FooterSegment::GitDiff {
-                        files_changed: stats.files_changed,
-                        additions: stats.total_additions,
-                        deletions: stats.total_deletions,
-                    })
-                }),
-                TuiStatuslineItem::CreditUsage => (!shell_mode)
-                    .then(|| self.selected_conversation_usage_totals(ctx))
-                    .flatten()
-                    .map(|totals| {
-                        let mode = AISettings::as_ref(ctx).usage_display_mode;
-                        FooterSegment::CreditUsage(self.usage_toggle.render_entry(
-                            mode,
-                            totals,
-                            ctx,
-                            |event_ctx, _| {
-                                event_ctx.dispatch_typed_action(
-                                    TuiTerminalSessionAction::ToggleUsageDisplay,
-                                );
-                            },
-                        ))
-                    }),
-                TuiStatuslineItem::ContextWindowUsage => (!shell_mode)
-                    .then(|| {
-                        self.conversation_selection
-                            .as_ref(ctx)
-                            .selected_conversation(ctx)
-                    })
-                    .flatten()
-                    .map(|conversation| {
-                        FooterSegment::ContextWindowUsage(format_context_window_usage(
-                            conversation.context_window_usage(),
-                        ))
-                    }),
-                TuiStatuslineItem::Date => Some(FooterSegment::DateTime(
-                    render_statusline_datetime(format_statusline_date, builder.muted_text_style()),
-                )),
-                TuiStatuslineItem::Time12Hour => {
-                    Some(FooterSegment::DateTime(render_statusline_datetime(
-                        format_statusline_time_12_hour,
-                        builder.muted_text_style(),
-                    )))
-                }
-                TuiStatuslineItem::Time24Hour => {
-                    Some(FooterSegment::DateTime(render_statusline_datetime(
-                        format_statusline_time_24_hour,
-                        builder.muted_text_style(),
-                    )))
-                }
-                TuiStatuslineItem::AgentTodoList => (!shell_mode)
-                    .then(|| {
-                        self.conversation_selection
-                            .as_ref(ctx)
-                            .selected_conversation(ctx)
-                    })
-                    .flatten()
-                    .and_then(|conversation| conversation.active_todo_list())
-                    .filter(|todo_list| !todo_list.is_empty())
-                    .map(|todo_list| {
-                        let hovered = self
-                            .todo_list_mouse
-                            .lock()
-                            .is_ok_and(|state| state.is_hovered());
-                        let style = if hovered {
-                            builder.primary_text_style()
-                        } else {
-                            builder.muted_text_style()
-                        };
-                        let progress = format_todo_progress(
-                            todo_list.completed_items().len(),
-                            todo_list.len(),
-                            todo_list.is_finished(),
-                        );
-                        FooterSegment::AgentTodoList(
-                            TuiHoverable::new(
-                                self.todo_list_mouse.clone(),
-                                TuiText::new(progress).with_style(style).truncate().finish(),
-                            )
-                            .on_click(|event_ctx, _| {
-                                event_ctx.dispatch_typed_action(
-                                    TuiTerminalSessionAction::ToggleTodoMenu,
-                                );
-                            })
-                            .finish(),
-                        )
-                    }),
-                TuiStatuslineItem::VoiceInput => voice_statusline_visible.then(|| {
-                    FooterSegment::VoiceInput(self.render_voice_statusline(&builder, ctx))
-                }),
-            };
-            if let Some(segment) = segment {
-                ordered.push(segment);
-            }
-        }
-        render_status_footer_row(FooterSegments { ordered }, &builder)
-    }
-
-    /// Returns a brief vim mode label for the footer when vim mode is enabled,
-    /// or `None` when vim mode is disabled.
-    fn vim_mode_indicator(&self, ctx: &AppContext) -> Option<&'static str> {
-        use vim::vim::{MotionType, VimMode};
-        let mode = self.input_view.as_ref(ctx).vim_mode(ctx)?;
-        match mode {
-            VimMode::Normal => Some("NOR"),
-            VimMode::Visual(MotionType::Charwise) => Some("VIS"),
-            VimMode::Visual(MotionType::Linewise) => Some("V-L"),
-            VimMode::Replace => Some("REP"),
-            // Insert mode is shown with a label, matching the GUI vim status indicator.
-            VimMode::Insert => Some("INS"),
-        }
-    }
-    fn render_auto_approve_statusline(
-        &self,
-        builder: &TuiUiBuilder,
-        ctx: &AppContext,
-    ) -> Box<dyn TuiElement> {
-        let enabled = self
-            .conversation_selection
-            .as_ref(ctx)
-            .pending_query_autoexecute_override(ctx)
-            .is_autoexecute_any_action();
-        let hovered = self
-            .footer_auto_approve_mouse
-            .lock()
-            .is_ok_and(|state| state.is_hovered());
-        let mut style = if enabled {
-            builder.success_glyph_style()
-        } else {
-            builder.muted_text_style()
-        };
-        if hovered {
-            style = style.add_modifier(Modifier::BOLD);
-        }
-        TuiHoverable::new(
-            self.footer_auto_approve_mouse.clone(),
-            TuiText::new("▶▶").with_style(style).truncate().finish(),
-        )
-        .on_click(|event_ctx, _| {
-            event_ctx.dispatch_typed_action(TuiTerminalSessionAction::ToggleAutoApprove {
-                show_feedback: false,
-            });
-        })
-        .finish()
-    }
-
-    fn voice_statusline_is_available(&self, shell_mode: bool, ctx: &AppContext) -> bool {
-        !shell_mode && AISettings::as_ref(ctx).is_voice_input_enabled(ctx)
-    }
-
-    fn render_voice_statusline(
-        &self,
-        builder: &TuiUiBuilder,
-        ctx: &AppContext,
-    ) -> Box<dyn TuiElement> {
-        let state = self.input_view.as_ref(ctx).voice_state(ctx);
-        let hovered = self
-            .voice_input_mouse
-            .lock()
-            .is_ok_and(|state| state.is_hovered());
-        let (label, style) = match state {
-            TuiVoiceInputState::Idle => (
-                "◉ Voice",
-                if hovered {
-                    builder.primary_text_style().add_modifier(Modifier::BOLD)
-                } else {
-                    builder.primary_text_style()
-                },
-            ),
-            TuiVoiceInputState::Listening => ("◉ Voice", builder.success_glyph_style()),
-            TuiVoiceInputState::Transcribing => {
-                return TuiText::new("… Transcribing")
-                    .with_style(builder.voice_input_status_style())
-                    .truncate()
-                    .finish();
-            }
-        };
-        TuiHoverable::new(
-            self.voice_input_mouse.clone(),
-            TuiText::new(label).with_style(style).truncate().finish(),
-        )
-        .on_click(|event_ctx, _| {
-            event_ctx
-                .dispatch_typed_action(TuiTerminalSessionAction::ToggleVoiceInputFromStatusline);
-        })
-        .finish()
-    }
-
-    /// Updates the watcher-backed git subscription after repository detection
-    /// completes for the active working directory. GitHub metadata is retained
-    /// only while its statusline item is enabled.
-    fn update_git_status_subscription(
-        &mut self,
-        repo_path: Option<LocalOrRemotePath>,
-        ctx: &mut ViewContext<Self>,
-    ) {
-        if self.current_repo_path == repo_path && self.git_repo_status.is_some() {
-            self.update_github_status_subscription(ctx);
-            return;
-        }
-        self.current_repo_path = repo_path.clone();
-        self.git_repo_status = None;
-        self.github_repo = None;
-
-        let Some(repo_path) = repo_path else {
-            ctx.notify();
-            return;
-        };
-        match GitRepoModels::handle(ctx)
-            .update(ctx, |models, ctx| models.subscribe(&repo_path, ctx))
-        {
-            Ok(handle) => {
-                ctx.subscribe_to_model(&handle, |_, _, _, ctx| ctx.notify());
-                self.git_repo_status = Some(handle);
-            }
-            Err(error) => {
-                log::warn!("Unable to subscribe TUI footer to git status: {error}");
-            }
-        }
-        self.update_github_status_subscription(ctx);
-        ctx.notify();
-    }
-
-    fn update_github_status_subscription(&mut self, ctx: &mut ViewContext<Self>) {
-        let enabled = AISettings::as_ref(ctx)
-            .tui_statusline
-            .normalized()
-            .is_enabled(TuiStatuslineItem::GitHubPullRequest);
-        if !enabled {
-            self.github_repo = None;
-            ctx.notify();
-            return;
-        }
-        if self.github_repo.is_some() {
-            return;
-        }
-        let Some(repo_path) = self.current_repo_path.clone() else {
-            return;
-        };
-        match GitRepoModels::handle(ctx).update(ctx, |models, ctx| {
-            models.subscribe_github_repo(&repo_path, ctx)
-        }) {
-            Ok(handle) => {
-                ctx.subscribe_to_model(&handle, |_, _, _, ctx| ctx.notify());
-                self.github_repo = Some(handle);
-            }
-            Err(error) => {
-                log::warn!("Unable to subscribe TUI footer to GitHub status: {error}");
-            }
-        }
-        ctx.notify();
-    }
-
-    fn git_status_metadata<'a>(&self, ctx: &'a AppContext) -> Option<&'a GitStatusMetadata> {
-        self.git_repo_status.as_ref()?.as_ref(ctx).metadata(ctx)
-    }
-
     /// Flips the footer usage entry's persisted credits⇄cost display mode.
     /// The settings-changed event re-renders every subscribed surface.
     fn toggle_usage_display(&mut self, ctx: &mut ViewContext<Self>) {
@@ -3956,20 +3231,6 @@ impl TuiTerminalSessionView {
                 menu.open(ctx);
             }
         });
-    }
-
-    /// The selected conversation's accumulated usage totals, or `None` (entry
-    /// hidden) until any usage has been reported.
-    fn selected_conversation_usage_totals(
-        &self,
-        ctx: &AppContext,
-    ) -> Option<ConversationUsageTotals> {
-        let totals = self
-            .conversation_selection
-            .as_ref(ctx)
-            .selected_conversation(ctx)?
-            .usage_totals();
-        (totals != ConversationUsageTotals::default()).then_some(totals)
     }
 
     /// The session's working directory. The cwd only arrives once shell
